refactor(tester): move bootstrap address resolution into KroxyliciousConfigUtils
Config traversal belongs in KroxyliciousConfigUtils alongside the other
config-building utilities. DefaultKroxyliciousTester now delegates to
KroxyliciousConfigUtils.bootstrapAddressFor(), injecting a port-resolver
BiFunction<Optional<String>, Integer, Integer> that calls
KafkaProxy.listeningPort() when the proxy is a KafkaProxy and falls
back to the identity for other proxy types.

The guard against OS_ASSIGNED_PORT is removed: the port resolver is
always called, which is a no-op for fixed ports and correct for
ephemeral ports.

KroxyliciousTester.listeningPort() is removed from the interface since
callers can obtain a bootstrap address directly via getBootstrapAddress().
